YQL-19747 Rename token.h to cursor_token_context.h
