+++ yql/essentials/sql/v1/complete/syntax/cursor_token_context_ut.cpp
@@ -1,4 +1,4 @@
-#include "token.h"
+#include "cursor_token_context.h"
 
 #include <library/cpp/testing/unittest/registar.h>
