+++ yql/essentials/sql/v1/complete/syntax/local.cpp
@@ -1,10 +1,10 @@
 #include "local.h"
 
 #include "ansi.h"
+#include "cursor_token_context.h"
 #include "format.h"
 #include "grammar.h"
 #include "parser_call_stack.h"
-#include "token.h"
 
 #include <yql/essentials/sql/v1/complete/antlr4/c3i.h>
 #include <yql/essentials/sql/v1/complete/antlr4/c3t.h>
