+++ yql/essentials/sql/v1/complete/syntax/ut/ya.make
@@ -2,7 +2,7 @@ UNITTEST_FOR(yql/essentials/sql/v1/complete/syntax)
 
 SRCS(
     grammar_ut.cpp
-    token_ut.cpp
+    cursor_token_context_ut.cpp
 )
 
 PEERDIR(
